// web/src/shared/api/generated/model/reportClientErrorRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ClientErrorSource } from './clientErrorSource';

export interface ReportClientErrorRequest {
  message: string;
  /** @nullable */
  errorName: string | null;
  /** @nullable */
  stack: string | null;
  source: ClientErrorSource;
  /** @nullable */
  path: string | null;
}